fix(link): ensure styles are applies when rendered within Popover in Menu
Fixes issue where Link styles are not applied when rendered within a PopoverContainer in a Menu.

fix #7762
